fixing expire and refund
